requirements.txt
python-lectio
python-dotenv